+++ tests/testthat/setup.R
@@ -1,0 +1,3 @@
+# avoid bug with cli and testthat
+# https://githubhot.com/repo/r-lib/cli/issues/441
+options(cli.hyperlink = FALSE)