+++ NEWS.md
@@ -1,3 +1,6 @@
+# hugostyle 1.0.2 _2022-04-22_
+  * test: render rmd against both system and rstudio pandoc versions
+
 # hugostyle 1.0.1 _2022-03-16_
   * Feat: Allow hugo snippets to pass through Rmd files